perf(init): stop compiling the same functions over and over
The 90s init budget is our tightest constraint -- rated round 98 used
88.5s of it -- so I profiled where it actually goes instead of guessing.
A listener on numba's compile events, charging each function only its own
time with nested compiles subtracted, says the import was 63 signatures
for 32 functions.

Two sources, both literal specialisation:

is_square_attacked took 6.2s across THIRTEEN compiles.  Twelve call sites
pass a constant castling square (E1, G8, ...) and numba specialises on
each literal value.  A pinned signature coerces them all to int64.

_search_root passed a bare `1` for ply.  That reaches numba as
Literal[int](1) and stays literal all the way down, forking _negamax and,
through it, _quiescence and _order_moves.  np.int64(1) de-literalises it
at the three root call sites, and the same for the `0` no-TT-move
sentinel into _order_moves.  allow_null stays literal deliberately: that
one really does let numba drop the null-move block from half the copies.

  signatures  63 -> 45
  is_square_attacked  13 -> 1
  _negamax             4 -> 2
  _order_moves         5 -> 2
  _quiescence          2 -> 1

Import measured 66.8s -> 56.2s, but both readings were taken with an
arena running, so treat the ratio rather than the seconds; the signature
counts are what is load-independent.  Scaled against round 98 that is
roughly 88.5s -> 74s.  tests.test_search_numba and tests.test_engine_numba
pass unchanged.

+++ current/search.py
@@ -754,8 +754,10 @@ def _quiescence(
         state,
         side,
         legal_stack[ply],
+        # np.int64(0), not 0 -- see the note in _search_root.  This is the
+        # "no TT move" sentinel and a bare 0 specialises _order_moves on it.
         count,
-        0,
+        np.int64(0),  # type: ignore[arg-type]
         ply,
         killers,
         quiet_history,
@@ -1507,6 +1509,16 @@ def _search_root(
         engine.make_move(pieces, state, key, move, undo_stack[0], undo_key_stack[0])
         history[history_count] = key[0]
         if index == 0:
+            # np.int64(1), not 1: a python int reaches numba as Literal[int](1)
+            # and stays literal the whole way down, so this one constant forked
+            # _negamax, and through it _quiescence and _order_moves, into extra
+            # compilations -- 63 signatures for 32 functions.  ply is an index
+            # into the ply-indexed stacks and specialising on its value buys
+            # nothing, while every extra signature is LLVM time inside the 90s
+            # init budget.  allow_null stays a literal on purpose: that one does
+            # let numba drop the null-move block from half the copies.
+            # mypy reads the dispatcher's annotation and wants a plain int;
+            # the value is identical, only its numba type differs.
             child_score, aborted = _negamax(
                 pieces,
                 state,
@@ -1514,7 +1526,7 @@ def _search_root(
                 depth - 1,
                 -beta,
                 -alpha,
-                1,
+                np.int64(1),  # type: ignore[arg-type]
                 True,
                 history,
                 history_count + 1,
@@ -1546,7 +1558,7 @@ def _search_root(
                 depth - 1,
                 -alpha - 1,
                 -alpha,
-                1,
+                np.int64(1),  # type: ignore[arg-type]
                 True,
                 history,
                 history_count + 1,
@@ -1578,7 +1590,7 @@ def _search_root(
                     depth - 1,
                     -beta,
                     -alpha,
-                    1,
+                    np.int64(1),  # type: ignore[arg-type]
                     True,
                     history,
                     history_count + 1,
